Avoid using pallet-executor in service by moving the required code to sp-executor
