+++ crates/pallet-executor/src/lib.rs
@@ -38,8 +38,8 @@ mod pallet {
     use frame_system::pallet_prelude::*;
     use sp_core::H256;
     use sp_executor::{
-        BundleEquivocationProof, ExecutionReceipt, ExecutorId, FraudProof, InvalidTransactionProof,
-        SignedExecutionReceipt, SignedOpaqueBundle,
+        BundleEquivocationProof, ExecutionReceipt, ExecutorId, FraudProof, InvalidTransactionCode,
+        InvalidTransactionProof, SignedExecutionReceipt, SignedOpaqueBundle,
     };
     use sp_runtime::traits::{
         CheckEqual, CheckedSub, MaybeDisplay, MaybeMallocSizeOf, One, SimpleBitOps,
@@ -392,27 +392,6 @@ mod pallet {
             // We need this extrinsic to be propagted to the farmer nodes.
             .propagate(true)
             .build()
-    }
-
-    #[repr(u8)]
-    pub enum InvalidTransactionCode {
-        BundleEquivicationProof = 101,
-        TrasactionProof = 102,
-        ExecutionReceipt = 103,
-        Bundle = 104,
-        FraudProof = 105,
-    }
-
-    impl From<InvalidTransactionCode> for InvalidTransaction {
-        fn from(invalid_code: InvalidTransactionCode) -> Self {
-            InvalidTransaction::Custom(invalid_code as u8)
-        }
-    }
-
-    impl From<InvalidTransactionCode> for TransactionValidity {
-        fn from(invalid_code: InvalidTransactionCode) -> Self {
-            InvalidTransaction::Custom(invalid_code as u8).into()
-        }
     }
 
     #[pallet::validate_unsigned]
